rename smart contract
