+++ contract/index.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-const MyAssetContract = require('./lib/my-asset-contract');
+const MyAssetContract = require('./lib/evoterContract.js');
 
 module.exports.MyAssetContract = MyAssetContract;
 module.exports.contracts = [ MyAssetContract ];
