+++ contract/lib/evoterContract.js
