+++ contract/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "contract",
+  "name": "evoterContract",
   "version": "0.0.1",
   "description": "My Smart Contract",
   "main": "index.js",
